Move PartInst to own file
